% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets-docs.R
\docType{data}
\encoding{UTF-8}
\name{aldist_example}
\alias{aldist_example}
\title{Age-length input data to gadget}
\format{
Data frame with gadget attributes containing Greenland halibut age-length distributions from the Institute of Marine Research EggaNord survey.
}
\source{
\href{https://www.hi.no/hi/nettrapporter/imr-pinro-en-2023-6}{Joint Russian-Norwegian Arctic Fisheries Working Group}
}
\usage{
aldist_example
}
\description{
Age-length input data to gadget
}
\seealso{
Other datasets: 
\code{\link{example_catches}},
\code{\link{fit}},
\code{\link{ldist_example}},
\code{\link{sexratio_example}},
\code{\link{stockdist_example}}
}
\concept{datasets}
\keyword{datasets}
\keyword{gadget.fit}
