% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internal_functions.R
\name{bind_fit_components}
\alias{bind_fit_components}
\title{Bind fit components over multiple gadget.fit objects}
\usage{
bind_fit_components(fit_list, component)
}
\arguments{
\item{fit_list}{A list of multiple gadget.fit objects}

\item{component}{Character specifying the list component in a gadget.fit object}
}
\description{
Bind fit components over multiple gadget.fit objects
}
\keyword{internal}
