% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dplot_adist.R
\name{dplot_adist}
\alias{dplot_adist}
\title{Plot age distribution data for a gadget3 model}
\usage{
dplot_adist(
  x,
  type = "bar",
  color_palette = scales::brewer_pal(palette = "Set1"),
  dir = "v",
  base_size = 8,
  ...
)
}
\arguments{
\item{x}{A gadget3 ready data frame created using mfdb, \link[gadgetutils]{g3_data} or \link[gadgetutils]{add_g3_attributes}.}

\item{type}{Character specifying the plot type: "bar" or "ggridges".}

\item{color_palette}{A function defining the color palette to be used for fill of bars.}

\item{dir}{Character specifying the direction of the facets ("h" for horizontal, "v" for vertical). See \code{\link[ggplot2]{facet_wrap}}. Defaults to "v" unlike in \code{\link[ggplot2]{facet_wrap}} to make it easier to follow length and age groups across years.}

\item{base_size}{Base size parameter for ggplot. See \link[ggplot2]{ggtheme}.}

\item{...}{Additional arguments passed to \code{\link[ggplot2]{facet_wrap}}.}
}
\value{
A \link[ggplot2]{ggplot} object.
}
\description{
The dplot functions plot data passed to a gadget3 model instead of data from the model or fit objects.
}
\examples{
data(aldist_example)
dplot_adist(aldist_example)
dplot_adist(aldist_example, type = "ggridges")
}
