% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dplot_aldist.R
\name{dplot_aldist}
\alias{dplot_aldist}
\title{Plot age-length distribution data for a gadget3 model}
\usage{
dplot_aldist(
  x,
  type = "bar",
  facet_age = FALSE,
  dir = "v",
  color_palette = scales::brewer_pal(palette = "Set1"),
  base_size = 8,
  ...
)
}
\arguments{
\item{x}{A gadget3 ready data frame created using mfdb, \link[gadgetutils]{g3_data} or \link[gadgetutils]{add_g3_attributes}.}

\item{type}{Character specifying the plot type: "bar", "step", or "area". "step" produces a similar plot where age and length distributions are plotted separately as in \link{plot_catchdist}. "area" produces an area plot instead of a bar plot, but works poorly when there are many age groups.}

\item{facet_age}{Logical indicating whether ages should be plotted in separate facets. Does not apply for \code{type = "step"}.}

\item{dir}{Character specifying the direction of the facets ("h" for horizontal, "v" for vertical). See \code{\link[ggplot2]{facet_wrap}}. Defaults to "v" unlike in \code{\link[ggplot2]{facet_wrap}} to make it easier to follow length and age groups across years.}

\item{color_palette}{A function defining the color palette to be used for fill of bars when \code{facet_age = TRUE}. See \link[ggplot2]{scale_color_manual}. To adjust color when \code{facet_age = FALSE}, use the standard \code{ggplot2::scale_fill_*} functions.}

\item{base_size}{Base size parameter for ggplot. See \link[ggplot2]{ggtheme}.}

\item{...}{Additional arguments passed to \code{\link[ggplot2]{facet_wrap}}.}
}
\value{
A \link[ggplot2]{ggplot} object.
}
\description{
The dplot functions plot data passed to a gadget3 model instead of data from the model or fit objects.
}
\examples{
data(aldist_example)
dplot_aldist(aldist_example)
dplot_aldist(aldist_example, type = "step") # as in plot_catchdist
dplot_aldist(aldist_example, facet_age = TRUE)
dplot_aldist(aldist_example, type = "area") # works poorly
}
