% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dplot_catch.R
\name{dplot_catch}
\alias{dplot_catch}
\title{Plot catch data for a gadget3 model}
\usage{
dplot_catch(x, base_size = 8)
}
\arguments{
\item{x}{A gadget3 ready data frame created using mfdb, \link[gadgetutils]{g3_data} or \link[gadgetutils]{add_g3_attributes}. Can be a single data frame or a named list of data frames in
which case multiple catches are plotted in the same figure.}

\item{base_size}{Base size parameter for ggplot. See \link[ggplot2]{ggtheme}.}
}
\value{
A \link[ggplot2]{ggplot} object.
}
\description{
The dplot functions plot data passed to a gadget3 model instead of data from the model or fit objects.
}
\examples{
data(example_catches)
dplot_catch(TrawlNor_catches)
dplot_catch(
  list("TrawlNor" = TrawlNor_catches,
       "OtherNor" = OtherNor_catches,
       "TrawlRus" = TrawlRus_catches,
       "OtherRus" = OtherRus_catches,
       "Internat" = Internat_catches)
       )
}
