% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dplot_ldist.R
\name{dplot_ldist}
\alias{dplot_ldist}
\title{Plot length distribution data for a gadget3 model}
\usage{
dplot_ldist(x, type = "bar", scales = "free_y", dir = "v", base_size = 8, ...)
}
\arguments{
\item{x}{A gadget3 ready data frame created using mfdb, \link[gadgetutils]{g3_data} or \link[gadgetutils]{add_g3_attributes}.}

\item{type}{Character specifying the plot type: "bar", "ggridges" or "line"}

\item{scales}{Character specifying the \code{scales} argument passed to \code{\link[ggplot2]{facet_wrap}}. Defaults to \code{"free_y"}.}

\item{dir}{Character specifying the direction of the facets ("h" for horizontal, "v" for vertical). See \code{\link[ggplot2]{facet_wrap}}. Defaults to "v" unlike in \code{\link[ggplot2]{facet_wrap}} to make it easier to follow length and age groups across years.}

\item{base_size}{Base size parameter for ggplot. See \link[ggplot2]{ggtheme}.}

\item{...}{Additional arguments passed to \code{\link[ggplot2]{facet_wrap}}.}
}
\value{
A \link[ggplot2]{ggplot} object.
}
\description{
The dplot functions plot data passed to a gadget3 model instead of data from the model or fit objects.
}
\examples{
data(ldist_example)
dplot_ldist(ldist_example)
dplot_ldist(ldist_example, type = "ggridges")
dplot_ldist(ldist_example, type = "line")
}
