% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dplot_stockdist.R
\name{dplot_stockdist}
\alias{dplot_stockdist}
\title{Plot stock distribution data for a gadget3 model}
\usage{
dplot_stockdist(
  x,
  stock_col = "maturity_stage",
  proportion = FALSE,
  group_by_sex = FALSE,
  sexes = c(female = "^female", male = "^male"),
  colors = NULL,
  scales = "free_y",
  dir = "v",
  base_size = 8,
  ...
)
}
\arguments{
\item{x}{A gadget3 ready data frame created using mfdb, \link[gadgetutils]{g3_data} or \link[gadgetutils]{add_g3_attributes}.}

\item{stock_col}{Character defining the name of the column separating stocks.}

\item{proportion}{Logical indicating whether to plot proportion of stocks instead
of absolute numbers.}

\item{group_by_sex}{Logical indicating whether to take the proportions by sex (\code{TRUE})
or by all stocks (\code{FALSE}). Used only when \code{proportion = TRUE}}

\item{sexes}{Named vector of length two containing regular expressions separating
sexes in \code{stock_col}. See the default as an example.}

\item{colors}{A vector of colors to be used for stocks. If \code{NULL},
\code{scales::hue_pal()} will be used}

\item{scales}{Character specifying the \code{scales} argument passed to \code{\link[ggplot2]{facet_wrap}}. Defaults to \code{"free_y"}.}

\item{dir}{Character specifying the direction of the facets ("h" for horizontal, "v" for vertical). See \code{\link[ggplot2]{facet_wrap}}. Defaults to "v" unlike in \code{\link[ggplot2]{facet_wrap}} to make it easier to follow length and age groups across years.}

\item{base_size}{Base size parameter for ggplot. See \link[ggplot2]{ggtheme}.}

\item{...}{Additional arguments passed to \code{\link[ggplot2]{facet_wrap}}.}
}
\value{
A \link[ggplot2]{ggplot} object.
}
\description{
The dplot functions plot data passed to a gadget3 model instead of data from the model or fit objects.
}
\examples{
data(stockdist_example)
dplot_stockdist(stockdist_example)
dplot_stockdist(stockdist_example, proportion = TRUE)
dplot_stockdist(stockdist_example, proportion = TRUE, group_by_sex = TRUE)
}
