% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets-docs.R
\docType{data}
\encoding{UTF-8}
\name{example_catches}
\alias{example_catches}
\alias{TrawlNor_catches}
\alias{Internat_catches}
\alias{OtherNor_catches}
\alias{OtherRus_catches}
\alias{TrawlRus_catches}
\title{Catch input data to gadget}
\format{
Data frame with gadget attributes containing Greenland halibut catches for ICES areas 1 and 2 (Northeast Arctic)

An object of class \code{data.frame} with 43 rows and 4 columns.

An object of class \code{data.frame} with 43 rows and 4 columns.

An object of class \code{data.frame} with 43 rows and 4 columns.

An object of class \code{data.frame} with 43 rows and 4 columns.

An object of class \code{data.frame} with 43 rows and 4 columns.

An object of class \code{data.frame} with 43 rows and 4 columns.
}
\source{
\href{https://www.hi.no/hi/nettrapporter/imr-pinro-en-2023-6}{Joint Russian-Norwegian Arctic Fisheries Working Group}
}
\usage{
TrawlNor_catches

Internat_catches

OtherNor_catches

Internat_catches

OtherRus_catches

TrawlRus_catches

TrawlNor_catches
}
\description{
Catch input data to gadget
}
\seealso{
Other datasets: 
\code{\link{aldist_example}},
\code{\link{fit}},
\code{\link{ldist_example}},
\code{\link{sexratio_example}},
\code{\link{stockdist_example}}
}
\concept{datasets}
\keyword{datasets}
\keyword{gadget.fit}
