% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets-docs.R
\docType{data}
\name{fit}
\alias{fit}
\title{An example gadget.fit object. Needs to be replaced by something better}
\format{
gadget.fit object.
}
\usage{
fit
}
\description{
An example gadget.fit object. Needs to be replaced by something better
}
\concept{datasets}
\keyword{datasets}
\keyword{gadget.fit}
