% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/g3d_plot.R
\name{g3d_plot}
\alias{g3d_plot}
\title{Plot of data passed into a gadget model}
\usage{
g3d_plot(model, regexp = NULL, scales = "fixed", ncol = NULL, base_size = 8)
}
\arguments{
\item{model}{R or TMB model. A model object from \link[gadget3]{g3_to_r} or \link[gadget3]{g3_to_tmb} functions.}

\item{regexp}{Either \code{NULL} (no filtering) or a character string giving a regular expression to filter model-data components. Useful alternatives: \code{"adist"} for abundance distribution data and \code{"cdist"} for catch distribution data.}

\item{scales}{Character defining the \code{\link[ggplot2]{facet_wrap}} \code{scales} argument to use.}

\item{ncol}{Number of columns passed to  \code{\link[ggplot2]{facet_wrap}}}

\item{base_size}{Base size parameter for ggplot. See \link[ggplot2]{ggtheme}.}
}
\value{
A list of \link[ggplot2]{ggplot} objects.
}
\description{
Plot of data passed into a gadget model
}
\details{
Plots data as contained in a gadget model. Helpful in checking models for possible data issues and for documentation of models.
}
