% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/g3plot_suitability.R
\name{g3plot_andersen}
\alias{g3plot_andersen}
\title{Plot Andersen suitability using model parameters}
\usage{
g3plot_andersen(
  length,
  p0,
  p1,
  p2,
  p3,
  p4,
  p5,
  add = FALSE,
  base_size = 8,
  ...
)
}
\arguments{
\item{length}{Numeric vector defining the lengths for which suitabilities should be plotted}

\item{p0, p1, p2, p3, p4, p5}{(Named) numeric vectors defining the Andersen suitability function parameters. See \link[gadget3]{suitability}. If length > 1, multiple curves will be plotted using a different line type. All of these parameters must have the same length.}

\item{add}{Logical indicating whether the function should produce a ggplot2 layer which can be added to an existing ggplot (\code{TRUE}) or whether to plot the results (\code{FALSE})}

\item{base_size}{Base size parameter for ggplot. See \link[ggplot2]{ggtheme}.}

\item{...}{Additional arguments passed to \code{plot_*} functions.}
}
\description{
Plot Andersen suitability using model parameters
}
\examples{
g3plot_andersen(length = 1:120, p0 = 0, p1 = 0.659, p2 = 1, p3 = 0.15, p4 = 1e4, p5 = 120)
}
