% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/g3plot_suitability.R
\name{g3plot_exponentiall50}
\alias{g3plot_exponentiall50}
\title{Plot exponentiall50 suitability using model parameters}
\usage{
g3plot_exponentiall50(length, alpha, l50, add = FALSE, base_size = 8, ...)
}
\arguments{
\item{length}{Numeric vector defining the lengths for which suitabilities should be plotted}

\item{alpha}{(Named) numeric vector defining the exponentiall50 slopes. If length > 1, multiple curves will be plotted using a different line type. See \link[gadget3]{suitability}.}

\item{l50}{Numeric vector defining the exponentiall50 l50 values. Must be same length than \code{alpha}}

\item{add}{Logical indicating whether the function should produce a ggplot2 layer which can be added to an existing ggplot (\code{TRUE}) or whether to plot the results (\code{FALSE})}

\item{base_size}{Base size parameter for ggplot. See \link[ggplot2]{ggtheme}.}

\item{...}{Additional arguments passed to \code{geom_line}}
}
\description{
Plot exponentiall50 suitability using model parameters
}
