% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gadget_plots.R
\name{gadget_plots}
\alias{gadget_plots}
\title{Wrapper for plot.gadget.fit that saves all diagnostic graphs to a directory}
\usage{
gadget_plots(
  fit,
  path,
  file_type = "png",
  quiet = FALSE,
  width = NULL,
  height = NULL,
  units = "cm",
  res = 300
)
}
\arguments{
\item{fit}{A gadget fit object. See \code{\link[gadgetutils]{g3_fit}}.}

\item{path}{Directory path for saving figures}

\item{file_type}{Character. Either one of the \code{device}s in \code{\link[ggplot2]{ggsave}}, in which case files of the defined type are printed to \code{path}, or "html" which compiles all plots into one html file using \code{\link[knitr]{knit}}.}

\item{quiet}{Logical indicating whether to print messages about the plotting process (set to \code{FALSE} to suppress the messages.}

\item{width, height}{Plot size given in \code{units}. If \code{NULL}, reasonable standard values are used.}

\item{units}{Units for plot size. See \code{\link[ggplot2]{ggsave}}}

\item{res}{Plot resolution. See the \code{dpi} argument in \code{\link[ggplot2]{ggsave}}}
}
\value{
Returns nothing, but makes the requested files.
}
\description{
Wrapper for plot.gadget.fit that saves all diagnostic graphs to a directory
}
