% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_html.R
\name{make_html}
\alias{make_html}
\title{Save diagnostic graphs to a single html document}
\usage{
make_html(
  fit,
  path = getwd(),
  harvest_rate = TRUE,
  file_name = "model_output_figures.html",
  template = "standard",
  model_name = "Gadget model output"
)
}
\arguments{
\item{fit}{A gadget fit object. See \code{\link[gadgetutils]{g3_fit}}.}

\item{path}{Directory path for saving the html output. Defaults to working directory.}

\item{harvest_rate}{A Logical value, if \code{TRUE} the harvest rate will be plotted, if \code{FALSE} the fishing mortality will be plotted.}

\item{file_name}{Character specifying the name of the html file without path. Must include the file extension.}

\item{template}{Character specifying the html template to use. See details.}

\item{model_name}{Character specifying the name of the model when "standard \code{template} is used.}
}
\value{
Returns nothing, but makes the requested file. See the \href{https://gadget-framework.github.io/gadgetplots/articles/make_html_output.html}{webpage for example output}.
}
\description{
Save diagnostic graphs to a single html document
}
\details{
The package contains html templates tailored for different stocks needed by the authors. Specify the name in the \code{template} argument. Current alternatives are:
\describe{
  \item{standard}{Standard model output figures. Use the \code{harvest_rate} argument to switch between F and HR figures.}
  \item{iceland}{Output tailored for Icelandic stock assessment.}
  \item{nea_ghl}{Output tailored for Northeast Arctic Greenland halibut assessment.}
}
}
