% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.gadget.fit.R
\name{plot.gadget.fit}
\alias{plot.gadget.fit}
\title{Plot \code{gadget.fit} object}
\usage{
\method{plot}{gadget.fit}(x, param = "annual", ...)
}
\arguments{
\item{x}{A gadget fit object. See \code{\link[gadgetutilss]{g3_fit}}.}

\item{param}{Character defining the parameter to plot. See Details.}

\item{...}{Additional parameters passed to the separate plotting functions.}
}
\value{
Single or a list of \link[ggplot2]{ggplot} objects depending on the arguments.
}
\description{
A wrapper function to plot the results from a \link[gadget3:g3_fit]{gadget.fit} object.
}
\details{
A wrapper to plot the results from a \link[gadget3:g3_fit]{gadget.fit} object. The function produces a different plots defined by \code{param} argument and additional arguments passed to the respective functions.

Valid \code{param} options are:
\describe{
  \item{annual}{Annual summary of central model data. Uses \code{\link{plot_annual}}}
  \item{sidat}{Surveyindices}
  \item{likelihood}{\link[=plot_likelihood]{Likelihood summary data}}
  \item{catchdist}{Catchdistribution data}
  \item{stockdist}{Stockdistribution data}
  \item{res.by.year}{Results by year}
  \item{stock.std}{Age composition from the model}
  \item{suitability}{Suitability estimated from the model by year and step}
  \item{params}{Parameter values relative to their boundaries}
}
and valid plottypes are:
\describe{
  \item{direct}{Default value, plots direct comparisons of data with model output. Valid for all datatype except 'res.by.year'}
  \item{weighted}{Only for 'summary'. Plots the weighted likelihood value for each component.}
  \item{pie}{Only for 'summary'. Plots the likelihood composition as a pie chart}
  \item{lengths}{Only for 'sidat'. Plot the surveyindex based on the SI length group instead of component name.}
  \item{bio}{Only for 'sidat'. Plot the biomass weighted survey index, assumes length based abundance index.}
  \item{scatter}{Only for 'sidat'. Produces a x-y scatter-plot for the fitted and observed index.}
  \item{resid}{Only for 'catchdist.fleets'. Produces a residual plot for each component.}
  \item{bubble}{Only for 'catchdist.fleets'. Produces a bubble plot for each component.}
  \item{growth}{Only for 'catchdist.fleets'. Produces a plot of fitted growth for each age-length component.}
  \item{resid}{Only for 'catchdist.fleets'. Produces a residual plot for each component.}
  \item{F}{Only for 'res.by.year'. Produces a F plot by stock.}
  \item{total}{Only for 'res.by.year'. Produces a total biomass plot by stock.}
  \item{catch}{Only for 'res.by.year'. Produces a total catch plot by stock.}
  \item{rec}{Only for 'res.by.year'. Produces a recruitment biomass plot by stock.}
}
}
