% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_adist.R
\name{plot_adist}
\alias{plot_adist}
\title{Plot model age distributions for stocks by year}
\usage{
plot_adist(
  fit,
  type = "bar",
  scales = "fixed",
  ncol = NULL,
  years = NULL,
  base_size = 8
)
}
\arguments{
\item{fit}{A gadget fit object. See \code{\link[gadgetutils]{g3_fit}}.}

\item{type}{Character specifying the plot type. Options: \code{"line"}, \code{"bar"} or \code{"ggridges"}. See Details.}

\item{scales}{Character defining the \code{\link[ggplot2]{facet_wrap}} \code{scales} argument to use.}

\item{ncol}{Number of columns passed to  \code{\link[ggplot2]{facet_wrap}}}

\item{years}{Numeric vector defining which years to plot. If \code{NULL} (default), all years are plotted.}

\item{base_size}{Base size parameter for ggplot. See \link[ggplot2]{ggtheme}.}
}
\value{
A \link[ggplot2]{ggplot} object or a list of such objects depending on the \code{type} argument.
}
\description{
Plot model age distributions for stocks by year
}
