% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_ageclass.R
\name{plot_ageclass}
\alias{plot_ageclass}
\title{Plot of age class from the model}
\usage{
plot_ageclass(
  fit,
  age = 2,
  biomass = FALSE,
  base_size = 8,
  return_data = FALSE
)
}
\arguments{
\item{fit}{A gadget fit object. See \code{\link[gadgetutils]{g3_fit}}.}

\item{age}{Integer giving the age class to plot.}

\item{biomass}{Logical indicating whether biomass should be plotted instead of estimated abundance.}

\item{base_size}{Base size parameter for ggplot. See \link[ggplot2]{ggtheme}.}

\item{return_data}{Logical indicating whether to return data for the plot instead of the plot itself.}
}
\value{
A \link[ggplot2]{ggplot} object.
}
\description{
Plot of age class from the model

Plots age class from the model at the beginning of the year
}
