% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_agelength.R
\name{plot_agelength}
\alias{plot_agelength}
\title{Plot of fitted growth for each age-length component}
\usage{
plot_agelength(fit, name = NULL, base_size = 8)
}
\arguments{
\item{fit}{A gadget fit object. See \code{\link[gadgetutils]{g3_fit}}.}

\item{name}{A character vector specifying the unique(fit$catchdist.fleets$name) to plot. If \code{NULL}, all names with relevant data are plotted.}

\item{base_size}{Base size parameter for ggplot. See \link[ggplot2]{ggtheme}.}
}
\value{
A \link[ggplot2]{ggplot} object. A list of ggplot objects if there are multiple age-length (aldist) data sources (\code{unique(fit$catchdist.fleets$name)}).
}
\description{
Plot of fitted growth for each age-length component
}
\examples{
data(fit)
plot_agelength(fit)
}
