% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_annual.R
\name{plot_annual}
\alias{plot_annual}
\title{Plot annual ICES type of graphic returning central model data}
\usage{
plot_annual(fit, base_size = 8, ...)
}
\arguments{
\item{fit}{A gadget fit object. See \code{\link[gadgetutils]{g3_fit}}.}

\item{base_size}{Base size parameter for ggplot. See \link[ggplot2]{ggtheme}.}

\item{...}{Additional arguments passed to \code{plot_*} functions. See Details.}
}
\value{
A \link[ggplot2]{ggplot} object. If \code{fleet = NULL}, a list of ggplot objects.
}
\description{
Plot annual ICES type of graphic returning central model data
}
