% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_biomass.R
\name{plot_biomass}
\alias{plot_biomass}
\title{Plot stock biomasses}
\usage{
plot_biomass(
  fit,
  type = "line",
  total = FALSE,
  biomass = TRUE,
  stocks = NULL,
  min_catch_length = NULL,
  return_data = FALSE,
  base_size = 8
)
}
\arguments{
\item{fit}{A gadget fit object. See \code{\link[gadgetutils]{g3_fit}}.}

\item{type}{Character specifying the plot type. Options: "line" or "area". See examples.}

\item{total}{Logical indicating whether total biomass should be plotted. Has no effect if \code{type = "area"}.}

\item{biomass}{Logical indicating whether biomass should be plotted instead of abundance.}

\item{stocks}{Character specifying the substock to plot in \code{fit}. If \code{NULL}, all stocks are plotted. Not applicable if \code{min_catch_length} is defined.}

\item{min_catch_length}{Numeric value defining the minimum catch length (size), which will be used to filter (\code{>=}) the model population before calculating biomass. Combines all stocks. Turn of by setting to \code{NULL} (default).}

\item{return_data}{Logical indicating whether to return data for the plot instead of the plot itself.}

\item{base_size}{Base size parameter for ggplot. See \link[ggplot2]{ggtheme}.}
}
\value{
A \link[ggplot2]{ggplot} object.
}
\description{
Plot stock biomasses
}
\examples{
data(fit)
plot_biomass(fit)
plot_biomass(fit, biomass = FALSE, total = TRUE)
plot_biomass(fit, type = "area")
plot_biomass(fit, min_catch_length = 45)
}
