% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_catch.R
\name{plot_catch}
\alias{plot_catch}
\title{Plot catches}
\usage{
plot_catch(fit, type = "stock", biomass = TRUE, base_size = 8)
}
\arguments{
\item{fit}{A gadget fit object. See \code{\link[gadgetutils]{g3_fit}}.}

\item{type}{Character specifying the data type: \code{"stock"} plots the catches by stock, \code{"fleet"} by fleet, and \code{"hr"} harvest rates by fleet.}

\item{biomass}{Logical indicating whether biomass should be plotted instead of estimated abundance.}

\item{base_size}{Base size parameter for ggplot. See \link[ggplot2]{ggtheme}.}
}
\value{
A \link[ggplot2]{ggplot} object.
}
\description{
Plot catches
}
