% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_catchdist.R
\name{plot_catchdist}
\alias{plot_catchdist}
\title{Plot catch distributions}
\usage{
plot_catchdist(fit, name = NULL, base_size = 8)
}
\arguments{
\item{fit}{A gadget fit object. See \code{\link[gadgetutils]{g3_fit}}.}

\item{name}{A character vector specifying the unique(fit$catchdist.fleets$name) to plot. If \code{NULL}, all names with relevant data are plotted.}

\item{base_size}{Base size parameter for ggplot. See \link[ggplot2]{ggtheme}.}
}
\value{
A list of \link[ggplot2]{ggplot} objects consisting of a separate plot for each of \code{names}.
}
\description{
Plots the length distribution of each \code{fit$catchdist.fleets$name}
}
