% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_f.R
\name{plot_f}
\alias{plot_f}
\title{Plot fishing mortality}
\usage{
plot_f(
  fit,
  stocks = NULL,
  fbar_ages = NULL,
  return_data = FALSE,
  base_size = 8
)
}
\arguments{
\item{fit}{A gadget fit object. See \code{\link[gadgetutils]{g3_fit}}.}

\item{stocks}{Character specifying the substock to plot in \code{fit}. If \code{NULL}, all stocks are plotted. Not applicable if \code{fbar_ages} is defined.}

\item{fbar_ages}{Either \code{NULL} or a numeric vector of ages to include to calculate Fbar (averaged F over age ranges) for selected ages instead of F for each stock.}

\item{return_data}{Logical indicating whether to return data for the plot instead of the plot itself. Used for jitter and retro plots.}

\item{base_size}{Base size parameter for ggplot. See \link[ggplot2]{ggtheme}.}
}
\value{
A \link[ggplot2]{ggplot} object.
}
\description{
Plot fishing mortality
}
\details{
The function calculates either average fishing mortality per substock or average fishing mortality over an age range depending on the \code{fbar_ages} argument.
}
\seealso{
plot_hr
}
