% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_growth.R
\name{plot_growth}
\alias{plot_growth}
\title{Plot of average growth for each stock}
\usage{
plot_growth(fit, type = "annual", stdev = FALSE, base_size = 8)
}
\arguments{
\item{fit}{A gadget fit object. See \code{\link[gadgetutils]{g3_fit}}.}

\item{type}{Character specifying the plot type. Options: \code{"annual"}, \code{"mean"} or \code{"stdev"}. See Details.}

\item{stdev}{Logical indicating whether standard deviation should be shown on both sides of mean for plots that use mean lengths}

\item{base_size}{Base size parameter for ggplot. See \link[ggplot2]{ggtheme}.}
}
\value{
A \link[ggplot2]{ggplot} object.
}
\description{
Plot of average growth for each stock
}
