% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_jitter.R
\name{plot_jitter}
\alias{plot_jitter}
\title{Plot central model output from a jitter analysis}
\usage{
plot_jitter(
  jitter_fit,
  vars = c("total.biomass", "hr", "recruitment"),
  base_size = 8,
  ncol = 1
)
}
\arguments{
\item{jitter_fit}{A list of \code{\link[gadgetutils]{g3_fit}} objects originating from \code{\link[gadgetutils]{g3_jitter}}.}

\item{vars}{Character vector giving the variables to plot. Only the default parameters are currently implemented.}

\item{base_size}{Base size parameter for ggplot. See \link[ggplot2]{ggtheme}.}

\item{ncol}{Number of columns to be used in the plot}
}
\value{
A \link[ggplot2]{ggplot} object.
}
\description{
Plot central model output from a jitter analysis
}
