% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_likelihood.R
\name{plot_likelihood}
\alias{plot_likelihood}
\title{Plot likelihood summary}
\usage{
plot_likelihood(
  fit,
  type = "total",
  log_scale = FALSE,
  use_proportions = TRUE,
  base_size = 8
)
}
\arguments{
\item{fit}{A gadget fit object. See \code{\link[gadgetutils]{g3_fit}}.}

\item{type}{Character specifying the plot type. Options: \code{"direct"}, \code{"weighted"} or \code{"total"}. See Details.}

\item{log_scale}{Logical indicating whether the value axis should be log10 transformed.}

\item{use_proportions}{Logical indicating whether proportions of summed likelihood scores should be used instead of summed likelihood score values when \code{type = "total"}.}

\item{base_size}{Base size parameter for ggplot. See \link[ggplot2]{ggtheme}.}
}
\value{
A \link[ggplot2]{ggplot} object.
}
\description{
Plot likelihood summary
}
\details{
Possible plot types are:
\describe{
  \item{direct}{Plots direct comparisons of data with model output.}
  \item{weighted}{Plots the weighted likelihood value for each component.}
  \item{total (or any other string)}{Plots the sums of weighted and raw likelihood values by component.}
  }
}
\examples{
data(fit)
plot_likelihood(fit)
plot_likelihood(fit, use_proportions = FALSE, log_scale = TRUE)
plot_likelihood(fit, type = "weighted")
}
