% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_param.R
\name{plot_param}
\alias{plot_param}
\title{Plot parameter values relative to their boundaries}
\usage{
plot_param(fit, out_only = FALSE, base_size = 8)
}
\arguments{
\item{fit}{A gadget fit object or a data frame containing model parameters. See \code{\link[gadgetutils]{g3_fit}}.}

\item{out_only}{Logical indicating whether only parameters outside their boundaries should be plotted.}

\item{base_size}{Base size parameter for ggplot. See \link[ggplot2]{ggtheme}.}
}
\value{
A \link[ggplot2]{ggplot} object.
}
\description{
Plot parameter values relative to their boundaries
}
\details{
The default plot is likely busy. Use \code{plotly::ggplotly()} to make it easier to interpret. If you are after parameters that are outside their boundaries, the \code{out_only} argument will help.
}
\examples{
data(fit)
plot_param(fit)
plot_param(fit, out_only = TRUE)
}
