% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_retro.R
\name{plot_retro}
\alias{plot_retro}
\title{Plot model biomass from a retrospective analysis}
\usage{
plot_retro(retro_fit, by_stock = TRUE, base_size = 8, ncol = NULL)
}
\arguments{
\item{retro_fit}{A list of \code{\link[gadgetutils]{g3_fit}} objects that differ in their terminal year, i.e. the result of a retrospective analysis. The first object should have no years removed and the objects should be sorted chronologically with increasing number of years removed.}

\item{by_stock}{Logical indicating whether to produce plots by stock in addition to total biomass.}

\item{base_size}{Base size parameter for ggplot. See \link[ggplot2]{ggtheme}.}

\item{ncol}{Number of columns to be used in the plot.}
}
\value{
A \link[ggplot2]{ggplot} object.
}
\description{
Plot model biomass from a retrospective analysis
}
