% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_si.R
\name{plot_si}
\alias{plot_si}
\title{Plot of survey indices}
\usage{
plot_si(fit, base_size = 8, type = "direct")
}
\arguments{
\item{fit}{A gadget fit object. See \code{\link[gadgetutils]{g3_fit}}.}

\item{base_size}{Base size parameter for ggplot. See \link[ggplot2]{ggtheme}.}

\item{type}{Character specifying the plot type: \code{"direct"} plots the model fit to survey data by year, and \code{"scatter"} produces a x-y scatter plot for the fitted and observed index.}
}
\value{
A \link[ggplot2]{ggplot} object.
}
\description{
Plot of survey indices
}
