% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_sr.R
\name{plot_sr}
\alias{plot_sr}
\title{Plot stock-recruitment relationship}
\usage{
plot_sr(
  fit,
  spawning_stock = NULL,
  lag = 1,
  years = NULL,
  add_line = TRUE,
  base_size = 8
)
}
\arguments{
\item{fit}{A gadget fit object. See \code{\link[gadgetutils]{g3_fit}}.}

\item{spawning_stock}{Either character defining the name of the stock that should be understood as spawning stock, or numeric giving the minimum length of spawning fish. If \code{NULL}, all stocks are used.}

\item{lag}{Integer giving the number of time steps that should be used as a lag between SSB and recruitment.}

\item{years}{A vector of recruitment years to include. If \code{NULL}, all years will be used.}

\item{add_line}{Add a line connecting years}

\item{base_size}{Base size parameter for ggplot. See \link[ggplot2]{ggtheme}.}
}
\value{
A \link[ggplot2]{ggplot} object or a list of such objects depending on the \code{stocks} argument.
}
\description{
Plots spawning stock biomass (SSB) against recruitment with a lag
}
\examples{
data(fit)
plot_sr(fit, spawning_stock = "ghl_female_mat")
plot_sr(fit, spawning_stock = 45)
}
