% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_ssb.R
\name{plot_ssb}
\alias{plot_ssb}
\title{Plot spawning stock biomass}
\usage{
plot_ssb(fit, spawning_stock = NULL, base_size = 8)
}
\arguments{
\item{fit}{A gadget fit object. See \code{\link[gadgetutils]{g3_fit}}.}

\item{spawning_stock}{Character specifying the name of the spawning stock in the \code{fit} object (i.e. the female mature stock). If \code{NULL}, the correct stock is guessed.}

\item{base_size}{Base size parameter for ggplot. See \link[ggplot2]{ggtheme}.}
}
\value{
A \link[ggplot2]{ggplot} object.
}
\description{
Plot spawning stock biomass
}
\examples{
data(fit)
plot_ssb(fit, spawning_stock = "ghl_female_mat")
}
