% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_stockcomp.R
\name{plot_stockcomp}
\alias{plot_stockcomp}
\title{Plot model stock composition}
\usage{
plot_stockcomp(fit, by_age = FALSE, geom_area = FALSE, base_size = 8)
}
\arguments{
\item{fit}{A gadget fit object. See \code{\link[gadgetutils]{g3_fit}}.}

\item{by_age}{Logical indicating whether age should be used on the x-axis instead of length}

\item{geom_area}{Logical indicating whether stacked area should be plotted instead of lines.}

\item{base_size}{Base size parameter for ggplot. See \link[ggplot2]{ggtheme}.}
}
\value{
A \link[ggplot2]{ggplot} object or a list of such objects depending on the \code{type} argument.
}
\description{
Plots proportions of stocks in the model by length or age
}
