% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_stockdist.R
\name{plot_stockdist}
\alias{plot_stockdist}
\title{Plot stock distribution}
\usage{
plot_stockdist(fit, stocks = NULL, base_size = 8)
}
\arguments{
\item{fit}{A gadget fit object. See \code{\link[gadgetutils]{g3_fit}}.}

\item{stocks}{Character vector specifying the stock to plot in \code{fit}. If \code{NULL}, all stocks are plotted in a single plot. If \code{"separate"}, all stocks are plotted in separate plots.}

\item{base_size}{Base size parameter for ggplot. See \link[ggplot2]{ggtheme}.}
}
\value{
A \link[ggplot2]{ggplot} object. A list of ggplot objects if \code{stocks = "separate"} or if there are multiple data sources (\code{unique(fit$stockdist$name)}).
}
\description{
Plots proportions of stocks by length or age in \code{fit$stockdist}
}
