% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_weight.R
\name{plot_weight}
\alias{plot_weight}
\title{Plot parameter weights}
\usage{
plot_weight(fit, log_scale = FALSE, base_size = 8)
}
\arguments{
\item{fit}{A gadget fit object. See \code{\link[gadgetutils]{g3_fit}}.}

\item{log_scale}{Logical indicating whether the value axis should be log10 transformed.}

\item{base_size}{Base size parameter for ggplot. See \link[ggplot2]{ggtheme}.}
}
\value{
A \link[ggplot2]{ggplot} object.
}
\description{
Plot parameter weights
}
\examples{
data(fit)
plot_weight(fit)
plot_weight(fit, log_scale = TRUE)
}
