% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/repeat_palette.R
\name{repeat_palette}
\alias{repeat_palette}
\title{Repeat colors from a color palette}
\usage{
repeat_palette(n, pal)
}
\arguments{
\item{n}{numeric defining how many times the length of levels in data}

\item{pal}{A function defining the color palette to be repeated.}
}
\value{
Returns a character vector of colors.
}
\description{
Repeats colors from a color palette \code{n} times to
ensure that vector of colors passed to ggplot2 manual color functions is as
long as the number of levels in data.
}
\examples{
scales::show_col(repeat_palette(12, scales::brewer_pal(palette = "Set1")))
}
\keyword{internal}
