% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets-docs.R
\docType{data}
\encoding{UTF-8}
\name{sexratio_example}
\alias{sexratio_example}
\title{Sex input data to gadget}
\format{
Data frame with gadget attributes containing Greenland halibut sex distributions (i.e. sex at length data) from the Barents Sea Ecosystem Survey.
}
\source{
\href{https://www.hi.no/hi/nettrapporter/imr-pinro-en-2023-6}{Joint Russian-Norwegian Arctic Fisheries Working Group}
}
\usage{
sexratio_example
}
\description{
Sex input data to gadget
}
\seealso{
Other datasets: 
\code{\link{aldist_example}},
\code{\link{example_catches}},
\code{\link{fit}},
\code{\link{ldist_example}},
\code{\link{stockdist_example}}
}
\concept{datasets}
\keyword{datasets}
\keyword{gadget.fit}
