% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets-docs.R
\docType{data}
\encoding{UTF-8}
\name{stockdist_example}
\alias{stockdist_example}
\title{Maturity input data to gadget}
\format{
Data frame with gadget attributes containing Greenland halibut stock distributions (i.e. maturity and sex at length data) from the Institute of Marine Research EggaNord survey.
}
\source{
\href{https://www.hi.no/hi/nettrapporter/imr-pinro-en-2023-6}{Joint Russian-Norwegian Arctic Fisheries Working Group}
}
\usage{
stockdist_example
}
\description{
Maturity input data to gadget
}
\seealso{
Other datasets: 
\code{\link{aldist_example}},
\code{\link{example_catches}},
\code{\link{fit}},
\code{\link{ldist_example}},
\code{\link{sexratio_example}}
}
\concept{datasets}
\keyword{datasets}
\keyword{gadget.fit}
