% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dplot_adist.R
\name{dplot_adist}
\alias{dplot_adist}
\title{Plot age distribution data for a gadget3 model}
\usage{
dplot_adist(
  x,
  type = "bar",
  color_palette = scales::brewer_pal(palette = "Set1"),
  scales = "fixed",
  ncol = NULL,
  base_size = 8
)
}
\arguments{
\item{x}{A gadget3 ready data frame created using mfdb, \link[gadgetutils]{g3_data} or \link[gadgetutils]{add_g3_attributes}.}

\item{type}{Character specifying the plot type: "bar" or "ggridges".}

\item{color_palette}{A function defining the color palette to be used for fill of bars.}

\item{scales}{Character defining the \code{\link[ggplot2]{facet_wrap}} \code{scales} argument to use.}

\item{ncol}{Number of columns passed to \code{\link[ggplot2]{facet_wrap}}}

\item{base_size}{Base size parameter for ggplot. See \link[ggplot2]{ggtheme}.}
}
\value{
A \link[ggplot2]{ggplot} object.
}
\description{
The dplot functions plot data passed to a gadget3 model instead of data from the model or fit objects.
}
\examples{
data(aldist_example)
dplot_adist(aldist_example)
dplot_adist(aldist_example, type = "ggridges")
}
