% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dplot_sexr.R
\name{dplot_sexr}
\alias{dplot_sexr}
\title{Plot sex ratio data for a gadget3 model}
\usage{
dplot_sexr(x, ncol = NULL, base_size = 8)
}
\arguments{
\item{x}{A gadget3 ready data frame created using mfdb, \link[gadgetutils]{g3_data} or \link[gadgetutils]{g3_data}.}

\item{ncol}{Number of columns passed to \code{\link[ggplot2]{facet_wrap}}}

\item{base_size}{Base size parameter for ggplot. See \link[ggplot2]{ggtheme}.}
}
\value{
A \link[ggplot2]{ggplot} object.
}
\description{
The dplot functions plot data passed to a gadget3 model instead of data from the model or fit objects.
}
