% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_html.R
\name{make_html}
\alias{make_html}
\title{Save diagnostic graphs to a single html document}
\usage{
make_html(fit, path, file_name = "model_output_figures.html")
}
\arguments{
\item{fit}{A gadget fit object. See \code{\link[gadgetutils]{g3_fit}}.}

\item{path}{Directory path for saving figures}

\item{file_name}{Character specifying the name of the html file without path. Must include the file extension.}
}
\value{
Returns nothing, but makes the requested file.
}
\description{
Save diagnostic graphs to a single html document
}
