% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.gadget.fit.R
\name{plot.gadget.fit}
\alias{plot.gadget.fit}
\title{Plot \code{gadget.fit} object}
\usage{
\method{plot}{gadget.fit}(x, param = "annual", ...)
}
\arguments{
\item{x}{A gadget fit object. See \code{\link[gadgetutilss]{g3_fit}}.}

\item{param}{Character defining the parameter to plot. See Details.}

\item{...}{Additional parameters passed to the separate plotting functions.}
}
\value{
Single or a list of \link[ggplot2]{ggplot} objects depending on the arguments.
}
\description{
A wrapper function to plot the results from a \link[gadget3:g3_fit]{gadget.fit} object.
}
\details{
A wrapper to plot the results from a \link[gadget3:g3_fit]{gadget.fit} object. The function produces a different plots defined by \code{param} argument and additional arguments passed to the respective functions.

Valid \code{param} options are:
\describe{
  \item{annual}{Annual summary of central model data. Uses \code{\link{plot_annual}}}
  \item{biomass}{Stock biomass. Uses \code{\link{plot_biomass}}}
  \item{total}{Stock biomass with total. Uses \code{\link{plot_biomass}}(..., total = TRUE)}
  \item{abundance}{Stock abundance. Uses \code{\link{plot_biomass}}(..., biomass = FALSE)}
  \item{ssb}{Spawning stock biomass. Uses \code{\link{plot_ssb}}}
  \item{rec}{Recruitment. Uses \code{\link{plot_rec}}}
  \item{f}{Fishing mortality. Uses \code{\link{plot_f}}}
  \item{catch}{Catches by stock. Uses \code{\link{plot_catch}}(..., type = "stock")}
  \item{catch}{Catches by fleet. Uses \code{\link{plot_catch}}(..., type = "fleet")}
  \item{hr}{Harvest rate. Uses \code{\link{plot_catch}}(..., type = "hr")}
  \item{si}{Survey indices. Uses \code{\link{plot_si}}}
  \item{catchdist}{Catch distribution. Uses \code{\link{plot_catchdist}}}
  \item{stockdist}{Stock distribution. Uses \code{\link{plot_stockdist}}}
  \item{suitablity}{Suitability (fleet selection). Uses \code{\link{plot_suitability}}}
  \item{growth}{Average length by age. Uses \code{\link{plot_growth}}}
  \item{agelength}{Age-length (growth parameter) fit. Uses \code{\link{plot_agelength}}}
  \item{agecomp}{Age composition. Uses \code{\link{plot_agecomp}}}
  \item{resid}{Residuals. Uses \code{\link{plot_resid}}}
  \item{likelihood}{Likelihood scores. Uses \code{\link{plot_likelihood}}}
  \item{weighted}{Weighted likelihood scores. Uses \code{\link{plot_likelihood}}(..., type = "weighted")}
  \item{pie}{Proportion of summed weighted likelihood scores. Uses \code{\link{plot_likelihood}}(..., type = "pie")}
  \item{params}{Parameter values relative to their boundaries (not implemented yet)}
}
}
