% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_agecomp.R
\name{plot_agecomp}
\alias{plot_agecomp}
\title{Plot of age composition from the model}
\usage{
plot_agecomp(fit, base_size = 8)
}
\arguments{
\item{fit}{A gadget fit object. See \code{\link[gadgetutils]{g3_fit}}.}

\item{base_size}{Base size parameter for ggplot. See \link[ggplot2]{ggtheme}.}
}
\value{
A \link[ggplot2]{ggplot} object.
}
\description{
Plot of age composition from the model
}
