% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_agecomp.R
\name{plot_agecomp}
\alias{plot_agecomp}
\title{Plot of age composition from the model}
\usage{
plot_agecomp(
  fit,
  type = "bubble",
  scales = "fixed",
  biomass = FALSE,
  base_size = 8,
  return_data = FALSE
)
}
\arguments{
\item{fit}{A gadget fit object. See \code{\link[gadgetutils]{g3_fit}}.}

\item{type}{Character specifying the plot type. Options: \code{"bar"}, \code{"line"}, \code{"ggridges"} or \code{"bubble"}. See Details.}

\item{scales}{Character defining the \code{\link[ggplot2]{facet_wrap}} \code{scales} argument to use.}

\item{biomass}{Logical indicating whether biomass should be plotted instead of estimated abundance.}

\item{base_size}{Base size parameter for ggplot. See \link[ggplot2]{ggtheme}.}

\item{return_data}{Logical indicating whether to return data for the plot instead of the plot itself.}
}
\value{
A \link[ggplot2]{ggplot} object.
}
\description{
Plot of age composition from the model
}
\details{
Possible plot types are:
\describe{
  \item{bar}{Facetted bar plot with year on x-axis, abundance on y-axis and ages along rows. Year classes are colored.}
  \item{line}{Facetted line plot with year on x-axis, abundance on y-axis and ages along rows. Stocks are colored.}
  \item{ggridges}{Same as above but the bars are overlapping.}
  \item{bubble (or any other string)}{Bubble plot with year on x-axis, age on y-axis and point size scaled to abundance. Year classes are colored.}
  }
Note that both \code{"bar"} and \code{"bubble"} contain extra information as text for \code{plotly::ggplotly}.
}
