% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_biomass.R
\name{plot_biomass}
\alias{plot_biomass}
\title{Plot stock biomasses}
\usage{
plot_biomass(
  fit,
  total = FALSE,
  geom_area = FALSE,
  biomass = TRUE,
  stocks = NULL,
  min_catch_length = NULL,
  return_data = FALSE,
  base_size = 8
)
}
\arguments{
\item{fit}{A gadget fit object. See \code{\link[gadgetutils]{g3_fit}}.}

\item{total}{Logical indicating whether total biomass should be plotted. Has no effect if \code{geom_area = TRUE}.}

\item{geom_area}{Logical indicating whether stacked area should be plotted instead of lines.}

\item{biomass}{Logical indicating whether biomass should be plotted instead of estimated abundance.}

\item{stocks}{Character specifying the substock to plot in \code{fit}. If \code{NULL}, all stocks are plotted. Not applicable if \code{min_catch_length} is defined.}

\item{min_catch_length}{Numeric value defining the minimum catch length (size), which will be used to filter (\code{>=}) the model population before calculating biomass. Combines all stocks. Turn of by setting to \code{NULL} (default).}

\item{return_data}{Logical indicating whether to return data for the plot instead of the plot itself. Used for jitter and retro plots.}

\item{base_size}{Base size parameter for ggplot. See \link[ggplot2]{ggtheme}.}
}
\value{
A \link[ggplot2]{ggplot} object.
}
\description{
Plot stock biomasses
}
