% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_f.R
\name{plot_f}
\alias{plot_f}
\title{Plot fishing mortality}
\usage{
plot_f(fit, stock = NULL, base_size = 8)
}
\arguments{
\item{fit}{A gadget fit object. See \code{\link[gadgetutils]{g3_fit}}.}

\item{stock}{Character specifying the stock to plot in \code{fit}. If \code{NULL}, all stocks are plotted.}

\item{base_size}{Base size parameter for ggplot. See \link[ggplot2]{ggtheme}.}
}
\value{
A \link[ggplot2]{ggplot} object.
}
\description{
Plot fishing mortality
}
\seealso{
plot_hr
}
