% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_growth.R
\name{plot_growth}
\alias{plot_growth}
\title{Plot of average growth for each stock}
\usage{
plot_growth(
  fit,
  type = "annual",
  stdev = FALSE,
  add_models = FALSE,
  base_size = 8
)
}
\arguments{
\item{fit}{A gadget fit object. See \code{\link[gadgetutils]{g3_fit}}.}

\item{type}{Character specifying the plot type. Options: \code{"annual"}, \code{"mean"} or \code{"stdev"}. See Details.}

\item{stdev}{Logical indicating whether standard deviation should be shown on both sides of mean for plots that use mean lengths.}

\item{add_models}{Logical indicating whether growth models extracted from growth parameters should be plotted together with the data. Uses grep and does not always work.}

\item{base_size}{Base size parameter for ggplot. See \link[ggplot2]{ggtheme}.}
}
\value{
A \link[ggplot2]{ggplot} object.
}
\description{
Plot of average growth for each stock
}
\examples{
data(fit)
plot_growth(fit)
plot_growth(fit, add_models = TRUE, stdev = TRUE)
plot_growth(fit, type = "mean")
plot_growth(fit, type = "stdev")
}
