% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_f.R
\name{plot_hr}
\alias{plot_hr}
\title{Plot harvest rate}
\usage{
plot_hr(
  fit,
  stock = NULL,
  min_catch_length = NULL,
  biomass = TRUE,
  base_size = 8
)
}
\arguments{
\item{fit}{A gadget fit object. See \code{\link[gadgetutils]{g3_fit}}.}

\item{stock}{Character specifying the stock to plot in \code{fit}. If \code{NULL}, all stocks are plotted. Not applicable if \code{min_catch_length} is defined.}

\item{min_catch_length}{Numeric value defining the minimum catch length (size), which will be used to filter (\code{>=}) the model population before calculating harvest rates using catches. Combines all stocks. Turn of by setting to \code{NULL} (default).}

\item{biomass}{Logical indicating whether biomass should be used to calculate harvest rates instead of abundance.}

\item{base_size}{Base size parameter for ggplot. See \link[ggplot2]{ggtheme}.}
}
\value{
A \link[ggplot2]{ggplot} object.
}
\description{
Plot harvest rate
}
\seealso{
plot_f
}
