% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_f.R
\name{plot_hr}
\alias{plot_hr}
\title{Plot harvest rate}
\usage{
plot_hr(
  fit,
  stocks = NULL,
  min_catch_length = NULL,
  biomass = TRUE,
  model_selectivity = TRUE,
  base_size = 8,
  return_data = FALSE
)
}
\arguments{
\item{fit}{A gadget fit object. See \code{\link[gadgetutils]{g3_fit}}.}

\item{stocks}{Character specifying the substock to plot in \code{fit}. If \code{NULL}, all stocks are plotted. Not applicable if \code{min_catch_length} is defined.}

\item{min_catch_length}{Numeric value defining the minimum catch length (size), which will be used to filter (\code{>=}) the model population before calculating harvest rates using catches. Combines all stocks. Turn of by setting to \code{NULL} (default). Set to 0 to get HR for the entire model population. See Details.}

\item{biomass}{Logical indicating whether biomass should be used to calculate harvest rates instead of abundance.}

\item{model_selectivity}{Logical indicating whether to use model selectivity when calculating harvest rates. \code{FALSE} will calculate harvest rate using the entire population, \code{TRUE} using the harvestable biomass following fleet suitabilities. Only applicable when using the \code{stocks} parameter or when \code{min_catch_length} = 0.}

\item{base_size}{Base size parameter for ggplot. See \link[ggplot2]{ggtheme}.}

\item{return_data}{Logical indicating whether to return data for the plot instead of the plot itself.}
}
\value{
A \link[ggplot2]{ggplot} object.
}
\description{
Plot harvest rate
}
\details{
The function uses all catches without filtering by size (but allocates them to substocks) and varies denominator depending on the \code{min_catch_length} argument. If \code{min_catch_length} = NULL, the output represents harvest rate for each substock assuming model fleet selectivities (fetched from the \code{harv.biomass} column in \code{fit$res.by.year}). If \code{min_catch_length} = 0, the output represents harvest rate for the entire model population assuming model fleet selectivities (fetched from the \code{harv.biomass} column in \code{fit$res.by.year}, stocks are summed before calculating the harvest rate). Finally, if \code{min_catch_length} is any number > 0, the output represents harvest rate \emph{assuming} flat selectivity for lengths >= \code{min_catch_length}. Note that the last \strong{IS NOT real model harvest rate}. It should only be used for reporting results, never for finding target harvest rates in projections, for instance. The \code{min_catch_length} = 0 is for the latter case.
}
\examples{
data(fit)
plot_hr(fit)
plot_hr(fit, model_selectivity = FALSE)
plot_hr(fit, min_catch_length = 45)
}
\seealso{
\link{plot_f}
}
