% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_ldist.R
\name{plot_ldist}
\alias{plot_ldist}
\title{Plot length distribution of model stocks by year}
\usage{
plot_ldist(fit, base_size = 8)
}
\arguments{
\item{fit}{A gadget fit object. See \code{\link[gadgetutils]{g3_fit}}.}

\item{base_size}{Base size parameter for ggplot. See \link[ggplot2]{ggtheme}.}
}
\value{
A \link[ggplot2]{ggplot} object or a list of such objects depending on the \code{type} argument.
}
\description{
Plot length distribution of model stocks by year
}
