% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_ldist.R
\name{plot_ldist}
\alias{plot_ldist}
\title{Plot model length distributions for stocks by year}
\usage{
plot_ldist(
  fit,
  type = "line",
  by_age = FALSE,
  stocks = NULL,
  scales = "fixed",
  ncol = NULL,
  years = NULL,
  base_size = 8
)
}
\arguments{
\item{fit}{A gadget fit object. See \code{\link[gadgetutils]{g3_fit}}.}

\item{type}{Character specifying the plot type. Options: \code{"line"}, \code{"bar"}, or \code{"ggridges"}.}

\item{by_age}{Logical indicating whether the length distributions should be grouped by age. Works only with \code{type = "line"}.}

\item{stocks}{Character specifying the substock to plot in \code{fit}. If \code{NULL}, all stocks are plotted.}

\item{scales}{Character defining the \code{\link[ggplot2]{facet_wrap}} \code{scales} argument to use.}

\item{ncol}{Number of columns passed to  \code{\link[ggplot2]{facet_wrap}}}

\item{years}{Numeric vector defining which years to plot. If \code{NULL} (default), all years are plotted.}

\item{base_size}{Base size parameter for ggplot. See \link[ggplot2]{ggtheme}.}
}
\value{
A \link[ggplot2]{ggplot} object or a list of such objects depending on the \code{type} argument.
}
\description{
Plot model length distributions for stocks by year
}
\details{
Do not trust the absolute numbers when \code{by_age = TRUE}. They are estimated from normal distributions by using \code{number * dnorm(1:120, mean = mean_length, sd = stddev_length)}.
}
\examples{
data(fit)
plot_ldist(fit)
plot_ldist(fit, type = "bar")
plot_ldist(fit, type = "ggridges")
plot_ldist(fit, by_age = TRUE, years = 2021)
}
