% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_leaveout.R
\name{plot_leaveout}
\alias{plot_leaveout}
\title{Plot central model output from a leave-out analysis}
\usage{
plot_leaveout(
  lo_fit,
  vars = c("nll.summary", "total.biomass", "hr", "recruitment"),
  base_size = 8,
  ncol = NULL
)
}
\arguments{
\item{lo_fit}{A list of \code{\link[gadgetutils]{g3_fit}} objects originating from a leave-out analysis. See details.}

\item{vars}{Character vector giving the variables to plot. Only the default parameters are currently implemented.}

\item{base_size}{Base size parameter for ggplot. See \link[ggplot2]{ggtheme}.}

\item{ncol}{Number of columns to be used in the plot. Set to \code{NULL} for standard layout.}
}
\value{
A \link[ggplot2]{ggplot} object.
}
\description{
Plot central model output from a leave-out analysis
}
\details{
Leave-out analysis has currently not been implemented to gadgetutils. Ask Mikko for the code.
}
