% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_stockcomp.R
\name{plot_stockcomp}
\alias{plot_stockcomp}
\title{Plot model stock composition}
\usage{
plot_stockcomp(fit, type = "line", by_age = FALSE, base_size = 8)
}
\arguments{
\item{fit}{A gadget fit object. See \code{\link[gadgetutils]{g3_fit}}.}

\item{type}{Character specifying the plot type. Options: "line" or "area". See examples.}

\item{by_age}{Logical indicating whether age should be used on the x-axis instead of length}

\item{base_size}{Base size parameter for ggplot. See \link[ggplot2]{ggtheme}.}
}
\value{
A \link[ggplot2]{ggplot} object or a list of such objects depending on the \code{type} argument.
}
\description{
Plots proportions of stocks in the model by length or age
}
\examples{
data(fit)
plot_stockcomp(fit)
plot_stockcomp(fit, type = "area")
plot_stockcomp(fit, by_age = TRUE)
}
