% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_stockdist.R
\name{plot_stockdist}
\alias{plot_stockdist}
\title{Plot stock distribution}
\usage{
plot_stockdist(
  fit,
  stocks = NULL,
  component_name = NULL,
  type = "model_fit",
  color_palette = scales::hue_pal(),
  geom_area = FALSE,
  base_size = 8
)
}
\arguments{
\item{fit}{A gadget fit object. See \code{\link[gadgetutils]{g3_fit}}.}

\item{stocks}{Character vector specifying the stock to plot in \code{fit}. If \code{NULL}, all stocks are plotted in a single plot. If \code{"separate"}, all stocks are plotted in separate plots.}

\item{component_name}{Character vector specifying the likelihood component to plot in \code{unique(fit$stockdist$name)}. If \code{NULL}, all components are plotted as a list. Only applicable for \code{type = "model_fit"}.}

\item{type}{Character specifying the plot type. Options: \code{"model_fit"}, \code{"stock_composition"} or \code{"sex_composition"}. See Details.}

\item{color_palette}{A function defining the color palette or a vector of colors to be used for stocks.}

\item{geom_area}{Logical indicating whether stacked area should be plotted instead of lines.}

\item{base_size}{Base size parameter for ggplot. See \link[ggplot2]{ggtheme}.}
}
\value{
A \link[ggplot2]{ggplot} object or a list of such objects depending on the \code{type} argument.
}
\description{
Plots proportions of stocks by length or age in \code{fit$stockdist}
}
\details{
Possible plot types are:
\describe{
  \item{model_fit}{Model fit to data using likelihoods. Separate plot for each \code{unique(fit$stockdist$name)}.}
  \item{stock_composition}{Overall model stock composition}
  \item{sex_composition}{Model sex composition. Not implemented yet.}
  }
}
